Frontend: Basic routing in App Component
Added some basic routing in the App Component to give the App some basic structure.

The Components Frontpage, Header, Layout, Login, BuildList, Charplanner and UsersList are mainly empty and will be finished at a later date. They were only added to be used in the routing inside the App Component.