Frontend: Create usePersist hook
Hook looks at the value of persist in local storage.

Creates a state called persist with the defaul value of false, if nothing was found in local storage. If persist exists in local storage it becomes the default of the persist state.

useEffect, which writes changes to persist immediately into the local storage.

hook returns both the persist state and the dispatch function.