Frontend: Fixed Charplanner component
When visiting a build directly through the link it would cause an error.

This was caused because the called hook changed between render.

Hooks in react shouldnt be used conditionally, therefore the code was rewritten to keep its functionality by moving the conditions inside of the callback functions in the useSelector hooks.