Frontend: Added updateBuild functionality to SaveBuild
Depending on the logged in user the function of the button changes. If it is the author of the build the button sends a patch request to update the build. Else it sends a post request to save the build.

This means another player can load someone elses build, do minor changes and then save it himself as a new build.