Frontend: Bugfix
When using validateStatus and invalidatesTags combined and you get an invalidStatus like 400 back, arg paramater for invalidatesTags will be undefined.

If you use the MutationTrigger inside a try catch block and you receive an invalid status back the error you will catch will be a type error because of arg being undefined.

This results in the error not having a status to check for, when setting a message to display to the user, what went wrong.

To fix this invalidatesTags will check, if arg exists. If it doesnt it returns an empty array.