Backend: Move validation of received user data from createNewUser to checkNewUserData
Move validation of received user data from createNewUser in usersController to middleware checkNewUserData.

This way a middleware can be executed after the validation of received user data and before the new user gets added to the database. This middleware is going to be a rateLimiter, which only allows 1 request per day. So a user can only add 1 new account per day.

This enables a user to send multiple request containing invalid user data. For example when a username or email is already in use, the user can send follow up requests. But once a user has send valid user data to create a new account the rateLimit will make sure he cant create more accounts after that.