Backend: Added sendreset and sendresetemail functions to authController
sendreset receives user in req.body and checks if user is allowed to reset password.

If yes a token for reset is created and saved into the database and sendresetemail is called with next.

sendresetemail takes the resetPasswordToken and the email out of the req and sends the token to said email using nodemailer.